backend-code/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1
SQLAlchemy==2.0.36
asyncpg==0.30.0
alembic==1.13.3
redis==5.0.8
minio==7.2.9
httpx==0.27.2
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
